feat: réintroduire le fichier navbar.tsx avec les sections et les icônes de navigation
